Improve extended bank data
Rename "BankData" to "ExtendedBankData" and explain its purpose

Move code that's specific to German IBANs into
KontoCheckBankDataGenerator, avoiding accidental calls of the code with
non-German IBANs and increasing coverage for the Iban class.
