+++ src/Domain/BankDataGenerator.php
@@ -4,7 +4,7 @@ declare( strict_types = 1 );
 
 namespace WMDE\Fundraising\PaymentContext\Domain;
 
-use WMDE\Fundraising\PaymentContext\Domain\Model\BankData;
+use WMDE\Fundraising\PaymentContext\Domain\Model\ExtendedBankData;
 use WMDE\Fundraising\PaymentContext\Domain\Model\Iban;
 
 /**
@@ -12,8 +12,8 @@ use WMDE\Fundraising\PaymentContext\Domain\Model\Iban;
  */
 interface BankDataGenerator {
 
-	public function getBankDataFromAccountData( string $account, string $bankCode ): BankData;
+	public function getBankDataFromAccountData( string $account, string $bankCode ): ExtendedBankData;
 
-	public function getBankDataFromIban( Iban $iban ): BankData;
+	public function getBankDataFromIban( Iban $iban ): ExtendedBankData;
 
 }
