+++ src/UseCases/GenerateBankData/BankDataSuccessResponse.php
@@ -4,12 +4,12 @@ declare( strict_types = 1 );
 
 namespace WMDE\Fundraising\PaymentContext\UseCases\GenerateBankData;
 
-use WMDE\Fundraising\PaymentContext\Domain\Model\BankData;
+use WMDE\Fundraising\PaymentContext\Domain\Model\ExtendedBankData;
 
 class BankDataSuccessResponse {
 
 	public function __construct(
-		public readonly BankData $bankData
+		public readonly ExtendedBankData $bankData
 	) {
 	}
 }
